Drop requirement for CA sources to use platform-specific suffix
This appears to be an arbitrary requirement with no practical
applications. Having it present means having to use workarounds
to add a `trusted_ca::ca` using an existing file on disk which
does not match the designated suffix for the platorm, or to
write cross-platform code.

This commit drops the requirement entirely, so that any source
can be used to install a trusted ca regardless of the filename.
Since the file is copied to the destination location using the
correct platform-specific suffix, there should be no
backward-compatibility breaks.

Fixes #38