Avoid running update-ca-trust twice on Red Hat
On both EL 7 & EL 8 the argument is ignored, so the command is
effectively running twice.

On EL 9 there is argument parsing. There is a difference how the enable
argument is handled. In ca-certificates-0:2024.2.69_v8.0.303-91.3.el9
it raises an error while release 91.4 changes it to a a deprecation
warning that users should use extract. Extract is almost the behavior
when no argument is given, but passing the argument is a bit stricter in
showing warnings and errors.

This simplifies the command to just plain update-ca-trust extract.